Add deserializer code
Add deserializer code for mediators without throwing synapse exceptions.
