+++ plugins/org.wso2.developerstudio.eclipse.gmf.esb.diagram/src/org/wso2/developerstudio/eclipse/gmf/esb/diagram/custom/deserializer/DummyAPIFactory.java
@@ -50,13 +50,16 @@ public class DummyAPIFactory {
 
 	public static API createAPI(OMElement apiElt, boolean withSynapse) {
 		OMAttribute nameAtt = apiElt.getAttribute(new QName("name"));
-		if (nameAtt == null || "".equals(nameAtt.getAttributeValue())) {
-			handleException("Attribute 'name' is required for an API definition");
-		}
-
 		OMAttribute contextAtt = apiElt.getAttribute(new QName("context"));
-		if (contextAtt == null || "".equals(contextAtt.getAttributeValue())) {
-			handleException("Attribute 'context' is required for an API definition");
+
+		if (withSynapse) {
+			if (nameAtt == null || "".equals(nameAtt.getAttributeValue())) {
+				handleException("Attribute 'name' is required for an API definition");
+			}
+
+			if (contextAtt == null || "".equals(contextAtt.getAttributeValue())) {
+				handleException("Attribute 'context' is required for an API definition");
+			}
 		}
 
 		API api = new API(nameAtt.getAttributeValue(), contextAtt.getAttributeValue());
